Fixed Continuous Bug
Fixed a bug in continuous wfc where some tiles would fail even though there were valid tile placement to be made. This was due to previous failed tiles would be added back to the open tile list and if they had the lowest entropy value then they are the one that would keep on getting selected for processing even though there are no other valid tile placement to be made. This mainly happened when tiles are fixed after processing. This bug has been fixed.

Also am showing the tile index value with the name in the UI. This will help in knowing which tiles are connected to each other much faster if the numbers are used instead.